resolve 3 months of merge conflicts and refactors
